Revert "Feature/picture and qr"
