+++ src/Routes.js
@@ -15,9 +15,6 @@ import MoreCounsel from './pages/QnA/MoreCounsel';
 import Prescription from './pages/Prescription/Prescription';
 import QnA from './pages/QnA/QnA';
 import Feed from './pages/Feed/Feed';
-import PictureAndQR from './pages/PictureAndQR/PictureAndQR';
-import SnapShot from './pages/SnapShot/SnapShot';
-import Loading from './pages/SnapShot/Loading';
 
 const Routes = () => {
   return (
@@ -37,11 +34,7 @@ const Routes = () => {
         <Route exact path="/prescription" component={Prescription} />
         <Route exact path="/qna" component={QnA} />
         <Route exact path="/feed" component={Feed} />
-        <Route exact path="/checkqr" component={PictureAndQR} />
-        <Route exact path="/snapshot" component={SnapShot} />
-        <Route exact path="/loading" component={Loading} />
       </Switch>
-      <div className="header_fixed" style={{ marginBottom: '69px' }} />
       <HomeBtn />
     </Router>
   );
