+++ src/components/HomeBtn.js
@@ -22,12 +22,13 @@ const HomeBtn = () => {
     </Home>
   );
 };
+
 export default HomeBtn;
 
 const Home = styled.div`
   width: 414px;
   height: 71px;
-  position: sticky;
+  position: fixed;
   bottom: 0;
   margin-top: -6px;
   background-color: black;
